chore: Moved elements to their own packages
