+++ go.mod
@@ -1,3 +1,3 @@
-module mediasms
+module github.com/comvex-jp/mediasms-go
 
 go 1.15
