+++ models/apiResponse.go
@@ -1,7 +1,7 @@
 package models
 
-// apiResponse struct
-type apiResponse struct {
+// APIResponse struct
+type APIResponse struct {
 	StatusCode  string
 	Name        string
 	Description string
