+++ translations/translations.go
@@ -1,0 +1,47 @@
+package translations
+
+// TranslationMap converts from Japanese to English
+var TranslationMap = map[string]string{
+	"失敗":              "Failure",
+	"送信済":             "Sent",
+	"送信中":             "Sending",
+	"リダイレクト":         "Redirect from shortened URL to target web sites",
+	"未送信":             "Not send yet",
+	"403":             "Mobile numbers are in the Black list",
+	"au 圏外・OFF":       "Error status from SDP",
+	"au 自 NW 障害":      "Error status from SDP",
+	"au 他 NW 障害":      "Error status from SDP",
+	"au 未払":           "Error status from SDP",
+	"処理中":             "Delivery Pending",
+	"au その他失敗":        "Error status for SDP",
+	"ドコモその他失敗":       "Error status for Docomo API",
+	"ドコモ圏外・OFF":      "Error status for Docomo API",
+	"ドコモ自 NW 障害":     "Error status for Docomo API",
+	"ドコモ他 NW 障害":     "Error status for Docomo API",
+	"無効":              "Invalid records",
+	"ソフトバンクその他失敗":    "Error status for Softbank API",
+	"ソフトバンク圏外・OFF":   "Error status for Softbank API",
+	"ソフトバンク自 NW 障害":  "Error status for Softbank API",
+	"ソフトバンク他 NW 障害":  "Error status for Softbank API",
+	"ソフトバンク受信拒否":     "Error status for Softbank API",
+	"送信不可":            "Error status for HLR API",
+	"送信前":             "SMS is scheduled, but not sent yet",
+	"一時停止":            "CSV file with SMS was paused",
+	"SMS 送信許可時間外":     "SMS was sent out of the allowed time",
+	"双方向 SMS 利用不可":    "No resourse for Reply",
+	"ドコモその他不明":       "Error status for Docomo API",
+	"重複不可":            "Duplicated sms",
+	"送信元重複":           "Keep chatID error",
+	"ドコモ一部受信":        "Error status for Docomo API",
+	"-":               "他人接続判定使用なし",
+	"キャンセル済":          "Cancelled SMS",
+	"送信上限":            "Ovelimitted SMS",
+	"NG ワード":         "Filtered SMS by stop words",
+	"配信停止希望":          "opted out",
+	"長文一部受信":          "Partial delivered via Gate",
+	"承認待ち":            "Wait for confirmation",
+	"Rakuten その他失敗":   "Error status for Rakuten",
+	"Rakuten 他 NW 障害": "Error status for Rakuten",
+	"Rakuten 一部受信":    "Error status for Rakuten",
+	"Rakuten 圏外・OFF":  "Error status for Rakuten",
+}
